# services/revuelta-api/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: ReVuelta REST API
  description: Operational Container Control System API for University Cafeteria Pilot V1
  version: 1.0.0
servers:
  - url: /api/v1
    description: Deployment-relative API base
paths:
...
  /circulations:
    post:
      summary: Register container delivery (create active circulation)
      description: Both signed QR payloads are revalidated atomically. A replay after a committed delivery returns `409 QR_ALREADY_USED`. No client idempotency key is accepted in the MVP.
      operationId: deliverContainer
      x-required-roles: [OPERATOR]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [participantQrPayload, containerQrPayload]
              properties:
                participantQrPayload: { type: string, minLength: 1, maxLength: 512 }
                containerQrPayload: { type: string, minLength: 1, maxLength: 512 }
      responses:
        '201':
          description: Atomic delivery result
          headers:
            X-Correlation-ID: { $ref: '#/components/headers/CorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Delivery' }
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthenticated'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
        '409':
          $ref: '#/components/responses/409Conflict'
        '500':
          $ref: '#/components/responses/500InternalServerError'
  /return-previews:
    post:
      summary: Validate both return QR values and obtain a read-only circulation preview
      description: Does not finalize the circulation, change container state or consume the participant QR.
      operationId: previewReturn
      x-required-roles: [OPERATOR]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DualQrHandoffRequest' }
      responses:
        '200':
          description: Current return eligibility and active circulation preview
          headers:
            X-Correlation-ID: { $ref: '#/components/headers/CorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReturnPreview' }
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthenticated'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
        '409':
          $ref: '#/components/responses/409Conflict'
        '500':
          $ref: '#/components/responses/500InternalServerError'
  /circulation-returns:
    post:
      summary: Register a physical container return using both QR values
      description: Both signed QR payloads are revalidated atomically. Success finalizes possession and leaves the container in `RETURNED` pending wash. A consumed participant QR receives `409 QR_ALREADY_USED`; a fresh QR for an already-returned container receives `409 RETURN_ALREADY_REGISTERED`.
      operationId: returnContainer
      x-required-roles: [OPERATOR]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DualQrHandoffRequest' }
      responses:
        '200':
          description: Atomic return result
          headers:
            X-Correlation-ID: { $ref: '#/components/headers/CorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReturnReceipt' }
        '400': { $ref: '#/components/responses/400BadRequest' }
        '401': { $ref: '#/components/responses/401Unauthenticated' }
        '403': { $ref: '#/components/responses/403Forbidden' }
        '404': { $ref: '#/components/responses/404NotFound' }
        '409': { $ref: '#/components/responses/409Conflict' }
        '500': { $ref: '#/components/responses/500InternalServerError' }
  /containers/{containerId}/history:
    get:
      summary: Get container audit history events
      operationId: getContainerEvents
      x-required-roles: [ADMIN]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: containerId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Append-oriented audit events for the container
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContainerEventPage' }
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthenticated'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '500':
          $ref: '#/components/responses/500InternalServerError'
  /me/circulations:
    get:
      summary: List circulations owned by the authenticated participant
      operationId: listMyCirculations
      x-required-roles: [PARTICIPANT]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [ACTIVE, COMPLETED] }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Participant-owned circulations
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ParticipantCirculationPage' }
        '400': { $ref: '#/components/responses/400BadRequest' }
        '401': { $ref: '#/components/responses/401Unauthenticated' }
        '403': { $ref: '#/components/responses/403Forbidden' }
        '409': { $ref: '#/components/responses/409Conflict' }
        '500': { $ref: '#/components/responses/500InternalServerError' }
  /me/circulations/{circulationId}:
    get:
      summary: Get one circulation owned by the authenticated participant
      operationId: getMyCirculation
      x-required-roles: [PARTICIPANT]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: circulationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Participant-owned circulation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ParticipantCirculation' }
        '400': { $ref: '#/components/responses/400BadRequest' }
        '401': { $ref: '#/components/responses/401Unauthenticated' }
        '403': { $ref: '#/components/responses/403Forbidden' }
        '404': { $ref: '#/components/responses/404NotFound' }
        '500': { $ref: '#/components/responses/500InternalServerError' }
...
  /operations/circulations:
    get:
      summary: List global circulations without participant PII
      operationId: listOperationsCirculations
      x-required-roles: [ADMIN]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [ACTIVE, COMPLETED] }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Global circulation records
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperationsCirculationPage' }
        '400': { $ref: '#/components/responses/400BadRequest' }
        '401': { $ref: '#/components/responses/401Unauthenticated' }
        '403': { $ref: '#/components/responses/403Forbidden' }
        '500': { $ref: '#/components/responses/500InternalServerError' }
...
components:
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  headers:
    CorrelationId:
      description: Server-generated identifier for correlating this response with operational logs
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProblemDetail:
      type: object
      required: [type, title, status, code, detail, instance, traceId, timestamp, errors]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string, example: "INVALID_CREDENTIALS" }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
        timestamp: { type: string, format: date-time }
        errors:
          type: array
          items: { type: string }
    ContainerState:
      type: string
      enum: [REGISTERED, AVAILABLE, IN_USE, RETURNED, DAMAGED, LOST, RETIRED]
...
    DeliveryContainer:
      type: object
      required: [id, publicCode, state]
      properties:
        id: { type: string, format: uuid }
        publicCode: { type: string }
        state: { type: string, enum: [AVAILABLE, IN_USE] }
    DualQrHandoffRequest:
      type: object
      required: [participantQrPayload, containerQrPayload]
      properties:
        participantQrPayload: { type: string, minLength: 1, maxLength: 512 }
        containerQrPayload: { type: string, minLength: 1, maxLength: 512 }
    DeliveryPolicy:
      type: object
      required: [id, version, name, durationHours]
      properties:
        id: { type: string, format: uuid }
        version: { type: integer, minimum: 1 }
        name: { type: string }
        durationHours: { type: integer, minimum: 1 }
...
    Delivery:
      type: object
      required: [circulationId, participantRef, container, deliveredAt, dueAt, policy, traceId]
      properties:
        circulationId: { type: string, format: uuid }
        participantRef: { type: string, format: uuid }
        container: { $ref: '#/components/schemas/DeliveryContainer' }
        deliveredAt: { type: string, format: date-time }
        dueAt: { type: string, format: date-time }
        policy: { $ref: '#/components/schemas/DeliveryPolicy' }
        traceId: { type: string, format: uuid }
    ReturnContainer:
      type: object
      required: [id, publicCode, state, stateLabel]
      properties:
        id: { type: string, format: uuid }
        publicCode: { type: string }
        state: { type: string, enum: [IN_USE, RETURNED] }
        stateLabel: { type: string, enum: [En uso, Pendiente de lavado] }
    ReturnPreview:
      type: object
      required: [circulationId, participantRef, container, deliveredAt, dueAt, previewedAt, traceId]
      properties:
        circulationId: { type: string, format: uuid }
        participantRef: { type: string, format: uuid }
        container: { $ref: '#/components/schemas/ReturnContainer' }
        deliveredAt: { type: string, format: date-time }
        dueAt: { type: string, format: date-time }
        previewedAt: { type: string, format: date-time }
        traceId: { type: string, format: uuid }
    ReturnReceipt:
      type: object
      required: [circulationId, participantRef, container, returnedAt, punctuality, traceId]
      properties:
        circulationId: { type: string, format: uuid }
        participantRef: { type: string, format: uuid }
        container: { $ref: '#/components/schemas/ReturnContainer' }
        returnedAt: { type: string, format: date-time }
        punctuality: { type: string, enum: [ON_TIME, LATE] }
        traceId: { type: string, format: uuid }
    ContainerEvent:
      type: object
      required: [id, containerId, eventType, actorId, occurredAt, newStatus, correlationId]
      properties:
        id: { type: string, format: uuid }
        containerId: { type: string, format: uuid }
        eventType: { type: string }
        actorId: { type: string, format: uuid }
        occurredAt: { type: string, format: date-time }
        previousStatus:
          type: string
          enum: [REGISTERED, AVAILABLE, IN_USE, RETURNED, DAMAGED, LOST, RETIRED]
          nullable: true
        newStatus:
          type: string
          enum: [REGISTERED, AVAILABLE, IN_USE, RETURNED, DAMAGED, LOST, RETIRED]
        reason: { type: string, nullable: true }
        correlationId: { type: string, format: uuid }
        participantRef: { type: string, format: uuid, nullable: true }
        circulationRef: { type: string, format: uuid, nullable: true }
    PageMetadata:
      type: object
      required: [page, size, hasNext]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
        hasNext: { type: boolean }
...
    ContainerEventPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          required: [items]
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/ContainerEvent' } }
    ParticipantCirculation:
      type: object
      required: [circulationId, containerId, publicCode, containerState, stateLabel, deliveredAt, dueAt, status]
      properties:
        circulationId: { type: string, format: uuid }
        containerId: { type: string, format: uuid }
        publicCode: { type: string }
        containerState: { $ref: '#/components/schemas/ContainerState' }
        stateLabel: { type: string }
        deliveredAt: { type: string, format: date-time }
        dueAt: { type: string, format: date-time }
        returnedAt: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [ACTIVE, COMPLETED] }
        punctuality: { type: string, enum: [ON_TIME, LATE], nullable: true }
    ParticipantCirculationPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          required: [items]
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/ParticipantCirculation' } }
...
    OperationsCirculation:
      type: object
      required: [circulationId, participantRef, containerId, publicCode, status, deliveredAt, dueAt]
      properties:
        circulationId: { type: string, format: uuid }
        participantRef: { type: string, format: uuid }
        containerId: { type: string, format: uuid }
        publicCode: { type: string }
        status: { type: string, enum: [ACTIVE, COMPLETED] }
        deliveredAt: { type: string, format: date-time }
        dueAt: { type: string, format: date-time }
        returnedAt: { type: string, format: date-time, nullable: true }
        punctuality: { type: string, enum: [ON_TIME, LATE], nullable: true }
    OperationsCirculationPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          required: [items]
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/OperationsCirculation' } }
...
  responses:
    401Unauthenticated:
      description: Authentication is missing, invalid or expired
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://revuelta.app/problems/unauthenticated
            title: Unauthorized
            status: 401
            code: UNAUTHENTICATED
            detail: Authentication is required or the supplied token is invalid
            instance: /api/v1/containers
            traceId: 3a06ba13-6a4c-49c1-92c7-13281f51d757
            timestamp: '2026-09-20T15:38:50Z'
            errors: []
    403Forbidden:
      description: The authenticated role is not authorized for the operation
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://revuelta.app/problems/forbidden-operation
            title: Forbidden
            status: 403
            code: FORBIDDEN_OPERATION
            detail: You are not authorized to perform this operation
            instance: /api/v1/containers
            traceId: ad0147fa-b2e0-4ad4-a751-dfadf4384a23
            timestamp: '2026-09-20T15:38:50Z'
            errors: []
...
    400BadRequest:
      description: Bad Request / Validation Failure
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    404NotFound:
      description: Resource Not Found
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    409Conflict:
      description: Business Conflict / State Machine Error
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    500InternalServerError:
      description: Unexpected internal failure; use traceId for support correlation
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'